split internals into submodules
