+++ foundations/src/telemetry/tokio_runtime_metrics/mod.rs
@@ -59,10 +59,12 @@
 //! }
 //! ```
 
-use crate::telemetry::metrics::{metrics, Counter, Gauge};
-use std::sync::atomic::Ordering;
+use crate::telemetry::tokio_runtime_metrics::runtime_handle::RuntimeHandle;
 use std::sync::Arc;
-use tokio::runtime::{Handle, RuntimeMetrics};
+use tokio::runtime::Handle;
+
+mod metrics;
+mod runtime_handle;
 
 /// Monitors a set of runtimes, allowing metrics to be periodically collected for each tracked runtime.
 #[derive(Default)]
@@ -114,251 +116,4 @@ impl RuntimeMonitor {
             .iter_mut()
             .for_each(RuntimeHandle::record_sample)
     }
-}
-
-struct RuntimeHandle {
-    runtime_name: Option<Arc<str>>,
-    runtime_id: Option<usize>,
-    handle: Handle,
-}
-
-impl RuntimeHandle {
-    fn record_sample(&mut self) {
-        let metrics = self.handle.metrics();
-
-        self.record_global_metrics(&metrics);
-
-        for worker_idx in 0..metrics.num_workers() {
-            self.record_worker_metrics(&metrics, worker_idx);
-        }
-    }
-
-    fn record_global_metrics(&mut self, metrics: &RuntimeMetrics) {
-        tokio_runtime::workers(&self.runtime_name, self.runtime_id)
-            .set(metrics.num_workers() as u64);
-
-        tokio_runtime::blocking_threads(&self.runtime_name, self.runtime_id)
-            .set(metrics.num_blocking_threads() as u64);
-
-        tokio_runtime::active_tasks(&self.runtime_name, self.runtime_id)
-            .set(metrics.active_tasks_count() as u64);
-
-        tokio_runtime::idle_blocking_threads(&self.runtime_name, self.runtime_id)
-            .set(metrics.num_idle_blocking_threads() as u64);
-
-        tokio_runtime::remote_schedules_total(&self.runtime_name, self.runtime_id)
-            .inner()
-            .store(metrics.remote_schedule_count(), Ordering::SeqCst);
-
-        tokio_runtime::budget_forced_yields_total(&self.runtime_name, self.runtime_id)
-            .inner()
-            .store(metrics.budget_forced_yield_count(), Ordering::SeqCst);
-
-        tokio_runtime::io_driver_fd_registrations_total(&self.runtime_name, self.runtime_id)
-            .inner()
-            .store(metrics.io_driver_fd_registered_count(), Ordering::SeqCst);
-
-        tokio_runtime::io_driver_fd_deregistrations_total(&self.runtime_name, self.runtime_id)
-            .inner()
-            .store(metrics.io_driver_fd_deregistered_count(), Ordering::SeqCst);
-
-        tokio_runtime::io_driver_fd_readies_total(&self.runtime_name, self.runtime_id)
-            .inner()
-            .store(metrics.io_driver_ready_count(), Ordering::SeqCst);
-
-        tokio_runtime::injection_queue_depth(&self.runtime_name, self.runtime_id)
-            .set(metrics.injection_queue_depth() as u64);
-
-        tokio_runtime::blocking_queue_depth(&self.runtime_name, self.runtime_id)
-            .set(metrics.blocking_queue_depth() as u64);
-    }
-
-    fn record_worker_metrics(&mut self, metrics: &RuntimeMetrics, worker_idx: usize) {
-        tokio_runtime_worker::parks_total(&self.runtime_name, self.runtime_id, worker_idx)
-            .inner()
-            .store(metrics.worker_park_count(worker_idx), Ordering::SeqCst);
-
-        tokio_runtime_worker::noops_total(&self.runtime_name, self.runtime_id, worker_idx)
-            .inner()
-            .store(metrics.worker_noop_count(worker_idx), Ordering::SeqCst);
-
-        tokio_runtime_worker::task_steals_total(&self.runtime_name, self.runtime_id, worker_idx)
-            .inner()
-            .store(metrics.worker_steal_count(worker_idx), Ordering::SeqCst);
-
-        tokio_runtime_worker::steal_operations_total(
-            &self.runtime_name,
-            self.runtime_id,
-            worker_idx,
-        )
-        .inner()
-        .store(
-            metrics.worker_steal_operations(worker_idx),
-            Ordering::SeqCst,
-        );
-
-        tokio_runtime_worker::polls_total(&self.runtime_name, self.runtime_id, worker_idx)
-            .inner()
-            .store(metrics.worker_poll_count(worker_idx), Ordering::SeqCst);
-
-        tokio_runtime_worker::busy_duration_micros_total(
-            &self.runtime_name,
-            self.runtime_id,
-            worker_idx,
-        )
-        .inner()
-        .store(
-            metrics.worker_total_busy_duration(worker_idx).as_micros() as u64,
-            Ordering::SeqCst,
-        );
-
-        tokio_runtime_worker::local_schedules_total(
-            &self.runtime_name,
-            self.runtime_id,
-            worker_idx,
-        )
-        .inner()
-        .store(
-            metrics.worker_local_schedule_count(worker_idx),
-            Ordering::SeqCst,
-        );
-
-        tokio_runtime_worker::overflows_total(&self.runtime_name, self.runtime_id, worker_idx)
-            .inner()
-            .store(metrics.worker_overflow_count(worker_idx), Ordering::SeqCst);
-
-        tokio_runtime_worker::local_queue_depth(&self.runtime_name, self.runtime_id, worker_idx)
-            .set(metrics.worker_local_queue_depth(worker_idx) as u64);
-
-        tokio_runtime_worker::mean_poll_time_micros(
-            &self.runtime_name,
-            self.runtime_id,
-            worker_idx,
-        )
-        .set(metrics.worker_mean_poll_time(worker_idx).as_micros() as u64);
-    }
-}
-
-#[metrics(crate_path = "crate")]
-mod tokio_runtime {
-    /// Number of worker threads in use by the runtime.
-    ///
-    /// This number shouldn't change during execution.
-    pub(super) fn workers(runtime_name: &Option<Arc<str>>, runtime_id: Option<usize>) -> Gauge;
-
-    /// Current number of blocking threads allocated by the runtime.
-    pub(super) fn blocking_threads(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Gauge;
-
-    /// Current number of active tasks on the runtime.
-    pub(super) fn active_tasks(runtime_name: &Option<Arc<str>>, runtime_id: Option<usize>)
-        -> Gauge;
-
-    /// Current number of idle blocking threads on the runtime which aren't doing anything.
-    pub(super) fn idle_blocking_threads(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Gauge;
-
-    /// Counter of schedules not originating from a worker on the runtime.
-    pub(super) fn remote_schedules_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Counter;
-
-    /// Counter of forced yields due to task budgeting.
-    pub(super) fn budget_forced_yields_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Counter;
-
-    pub(super) fn io_driver_fd_registrations_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Counter;
-
-    pub(super) fn io_driver_fd_deregistrations_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Counter;
-
-    pub(super) fn io_driver_fd_readies_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Counter;
-
-    pub(super) fn injection_queue_depth(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Gauge;
-
-    pub(super) fn blocking_queue_depth(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-    ) -> Gauge;
-}
-
-#[metrics(crate_path = "crate")]
-mod tokio_runtime_worker {
-    pub(super) fn parks_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn noops_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn task_steals_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn steal_operations_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn polls_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn busy_duration_micros_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn local_schedules_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn overflows_total(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Counter;
-
-    pub(super) fn local_queue_depth(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Gauge;
-
-    pub(super) fn mean_poll_time_micros(
-        runtime_name: &Option<Arc<str>>,
-        runtime_id: Option<usize>,
-        worker_idx: usize,
-    ) -> Gauge;
 }
